DABOM-146 refactor: Consumer 책임 축소, Validator/SyncService로 분리
